Fix: update gemfile with new addressable module
