+++ config/application.rb
@@ -8,6 +8,7 @@ Bundler.require(*Rails.groups)
 
 module HackathonApi
   class Application < Rails::Application
+    config.autoload_paths += %W(#{config.root}/lib)
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
